Edited UI

+++ interest/uap/views.py
@@ -10,9 +10,11 @@ from decimal import Decimal
 
 def index(request):
     # Query the database to get the required information
-    total_deposit = Investment.objects.aggregate(Sum('deposit')).get('deposit__sum', 0)
-    interest = Decimal('0.00324544') * total_deposit  # 10% interest on total_deposit
-    running_investments = total_deposit + interest
+    total_deposit = int(Investment.objects.aggregate(Sum('deposit')).get('deposit__sum', 0))
+    interest = int(Decimal('0.00324544') * total_deposit) # 10% interest on total_deposit
+    running_investments = int(total_deposit + interest)
+
+
    
     # Pass the data to the template for rendering
     return render(request, 'index.html', {
